added .toFloat to the division resulting in neededProbability, to store the correct float value (instead of all zeros!)

+++ src/main/scala-2.10/RandomSamplingReservoir.scala
@@ -57,7 +57,7 @@ object RandomSamplingReservoir {
 
           //do probabilities here :D
           randomProbability = randomGenerator.nextFloat()
-          neededProbabilityForInclusion = reservoirSize/(elementCount+1)
+          neededProbabilityForInclusion = reservoirSize.toFloat/(elementCount+1)
 
           println("NEEEEDED " + neededProbabilityForInclusion)
